refactor: fine-tuning
ref #2318
