+++ visualization/app/codeCharta/ui/ribbonBar/artificialIntelligence/highRiskProfile/highRiskProfile.component.spec.ts
@@ -1,12 +1,13 @@
 import { render, screen } from "@testing-library/angular"
 import userEvent from "@testing-library/user-event"
-import { MaterialModule } from "../../../../../material/material.module"
+import { ArtificialIntelligenceModule } from "../artificialIntelligence.module"
 import { HighRiskProfileComponent } from "./highRiskProfile.component"
 
 describe("HighRiskProfileComponent", () => {
 	it("should render profile with an analyzed programming properties", async () => {
 		await render(HighRiskProfileComponent, {
-			imports: [MaterialModule],
+			excludeComponentDeclaration: true,
+			imports: [ArtificialIntelligenceModule],
 			componentProperties: {
 				data: {
 					analyzedProgrammingLanguage: "ts",
@@ -44,7 +45,8 @@ describe("HighRiskProfileComponent", () => {
 
 	it("should explain that there couldn't be analyzed a risk profile", async () => {
 		await render(HighRiskProfileComponent, {
-			imports: [MaterialModule],
+			excludeComponentDeclaration: true,
+			imports: [ArtificialIntelligenceModule],
 			componentProperties: {
 				data: {
 					analyzedProgrammingLanguage: "",
