+++ visualization/app/codeCharta/ui/ribbonBar/artificialIntelligence/suspiciousMetrics/suspiciousMetrics.component.spec.ts
@@ -1,13 +1,18 @@
 import { TestBed } from "@angular/core/testing"
 import { render, screen } from "@testing-library/angular"
 import userEvent from "@testing-library/user-event"
+import { Store } from "../../../../state/angular-redux/store"
+import { setColorMetric } from "../../../../state/store/dynamicSettings/colorMetric/colorMetric.actions"
+import { setColorRange } from "../../../../state/store/dynamicSettings/colorRange/colorRange.actions"
+import { setHeightMetric } from "../../../../state/store/dynamicSettings/heightMetric/heightMetric.actions"
 import { ArtificialIntelligenceModule } from "../artificialIntelligence.module"
 import { SuspiciousMetricComponent } from "./suspiciousMetrics.component"
 
 describe("SuspiciousMetricsComponent", () => {
 	beforeEach(() => {
 		TestBed.configureTestingModule({
-			imports: [ArtificialIntelligenceModule]
+			imports: [ArtificialIntelligenceModule],
+			providers: [{ provide: Store, useValue: { dispatch: jest.fn() } }]
 		})
 	})
 
@@ -79,6 +84,12 @@ describe("SuspiciousMetricsComponent", () => {
 			expect(screen.getByText("rloc")).not.toBe(null)
 			expect(screen.getByText("Suspicious Metrics in ts code")).not.toBe(null)
 			expect(screen.getByText("Suspicious MCC Files")).not.toBe(null)
+
+			userEvent.click(screen.getByText("Suspicious MCC Files"), undefined, { skipPointerEventsCheck: true })
+			const store = TestBed.inject(Store)
+			expect(store.dispatch).toHaveBeenCalledWith(setHeightMetric("mcc"))
+			expect(store.dispatch).toHaveBeenCalledWith(setColorMetric("mcc"))
+			expect(store.dispatch).toHaveBeenCalledWith(setColorRange({ from: 10, to: 22 }))
 		})
 	})
 })
